feat: #27 SelectSpotCategoryFeature 이전
